Meeting Scheduling JSON file
This is a JSON file that contains the code for automating the process that is discussed in the ReadMe file. This should be imported into the n8n workflow, and credential matching must be done in order to use this file to its full extent. Read the readme file to understand how to import and set credentials.